[generator] Aligned generator templates with the latest SDK changes for `IndistinctAsset` fallback model

+++ pyatlan/generator/class_generator.py
@@ -938,6 +938,9 @@ if __name__ == "__main__":
     for file in (ASSETS_DIR).glob("*.py"):
         file.unlink()
     for file in (CORE_ASSETS_DIR).glob("*.py"):
+        # Remove all files in the core assets directory except `indistinct.py`
+        if file.name == "indistinct_asset.py":
+            continue
         file.unlink()
     generator = Generator()
     EnumDefInfo.create(type_defs.enum_defs)
